+++ file_reader/safe_reader.py
@@ -1,16 +1,17 @@
-from file_reader.registry import get
-import tempfile
 from pathlib import Path
 import shutil as sh
+import tempfile
+
 import pandas as pd
 
+from file_reader import registry
 
 class SafeReader():
     def __init__(self, file_path: str):
         self.temp_dir = tempfile.mkdtemp()
         self.file_path = self._temp_path(file_path)
-        self.reader_func = get(Path(self.file_path).suffix)
-        self.data: pd.DataFrame = self.reader_func(file_path)
+        self.reader_func = registry.READERS_REGISTRY[Path(self.file_path).suffix]
+        self.data: pd.DataFrame = self.reader_func(self.file_path)
         
     def __enter__(self):
         return self
@@ -25,24 +26,27 @@ class SafeReader():
         else:
             raise ValueError(f"{file_path} is not a valid file path")
 
-    def _confirm_path(self, file_path):
+    def _confirm_path(self, file_path: str|Path) -> Path:
         
-        # is it a string?
-        if not isinstance(file_path, str):
-            raise TypeError(f"_confirm_path: expected type str, recieved: {type(self.file_path)}")
+        # is it a string?        
+        if not isinstance(file_path, (str, Path)):
+            raise TypeError(f"_confirm_path: expected str|Path, received: {type(file_path)}")
         
         # if string, is it a valid file path?
-        try:
+        if isinstance(file_path, str):
+            try:
+                fp = Path(file_path)
+            except Exception as e:
+                raise ValueError(f"_confirm_path: invalid file path, {file_path}")
+        else:
             fp = Path(file_path)
-        except Exception as e:
-            raise ValueError(f"_confirm_path: invalid file path, {file_path}")
-        
+
         # does the file exist?
         if not fp.exists():
-            raise FileExistsError(f"_confirm_path: {fp} does not exist")
+            raise FileNotFoundError(f"_confirm_path: {fp} does not exist")
         
         # passed all tests, return Path object
-        return Path(file_path)
+        return file_path
 
     def _copy_to_temp(self, file_path) -> Path:
         dst_path = Path(self.temp_dir) / Path(file_path).name