+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### [Unreleased]
-- item one
+- convert entire directory to single data frame
 - item two
 - item three